backend/requirements.txt
# Django
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
django-environ==0.11.2

# Database
psycopg2-binary==2.9.10

# Redis & Celery
redis==5.2.1
celery==5.4.0
django-redis==5.4.0
django-celery-beat==2.7.0

# Auth & Security
djangorestframework-simplejwt==5.4.0
bcrypt==4.2.1
cryptography==44.0.0

# Google OAuth
google-auth==2.37.0

# HTTP
requests==2.32.3
httpx==0.28.1

# Crypto & Blockchain
tronpy==0.5.0
web3==7.6.0
bitcoinlib==0.7.7
mnemonic==0.21

# Rate feeds
pycoingecko==3.2.0

# SMS
africastalking==1.2.7

# Production
gunicorn==23.0.0
whitenoise==6.8.2
sentry-sdk[django]==2.19.2

# Dev & Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-asyncio==0.24.0
factory-boy==3.3.1
faker==33.1.0
